@@ -1,7 +1,7 @@
 {
   "name": "tns-core-modules",
   "description": "Telerik NativeScript Core Modules",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "homepage":"https://www.nativescript.org",
   "repository": {
     "type": "git",
@@ -31,10 +31,10 @@
     "grunt-untar": "0.0.1",
     "mocha": "2.2.5",
     "shelljs": "0.5.3",
-    "grunt-tslint": "3.0.0",
+    "grunt-tslint": "3.0.3",
     "time-grunt": "^1.3.0",
-    "tslint": "3.3.0",
-    "typescript": "1.7.5"
+    "tslint": "3.4.0",
+    "typescript": "1.8.2"
   },
   "typings": "tns-core-modules.d.ts",
   "dependencies": {
