@@ -1,7 +1,6 @@
-﻿import application = require("application");
-import navPageModule = require("../nav-page");
-
-import trace = require("trace");
+﻿import * as application from "application";
+import {NavPage} from "../nav-page";
+import * as trace from "trace";
 trace.enable();
 trace.setCategories(trace.categories.concat(
     trace.categories.NativeLifecycle,
@@ -11,18 +10,14 @@ trace.setCategories(trace.categories.concat(
 ));
 
 application.mainEntry = {
-    create: function () {
-        return new navPageModule.NavPage({
-            index: 0,
-            backStackVisible: true,
-            clearHistory: false,
-            animated: true,
-            transition: 0,
-            curve: 0,
-            duration: 0,
-        });
-    }
-    //backstackVisible: false,
-    //clearHistory: true
+    create: () => new NavPage({
+        index: 0,
+        backStackVisible: true,
+        clearHistory: false,
+        animated: true,
+        transition: 0,
+        curve: 0,
+        duration: 0,
+    })
 };
 application.start();
