@@ -1,18 +1,19 @@
-﻿import TKUnit = require("../../TKUnit");
-import helper = require("../helper");
-import viewModule = require("ui/core/view");
-import observable = require("data/observable");
-import color = require("color");
-import platform = require("platform");
-
+﻿import * as TKUnit from "../../TKUnit";
+import * as helper from "../helper";
+import {View} from "ui/core/view";
+import {Button} from "ui/button";
+import {Page} from "ui/page";
+import {ScrollView} from "ui/scroll-view";
+import {LayoutBase} from "ui/layouts/layout-base"
+import {StackLayout} from "ui/layouts/stack-layout"
+import {GridLayout} from "ui/layouts/grid-layout"
 import {ProxyViewContainer} from "ui/proxy-view-container";
-import {View, Button, LayoutBase, StackLayout, GridLayout, Page, ScrollView, TabView, TabViewItem} from "ui";
 
 export function test_add_children_to_attached_proxy() {
     var outer = new StackLayout();
     var proxy = new ProxyViewContainer();
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         outer.addChild(createBtn("1"));
         outer.addChild(proxy);
         proxy.addChild(createBtn("2"));
@@ -31,7 +32,7 @@ export function test_children_immediately_registered_in_parent_grid_layout() {
     var outer = new GridLayout();
     var proxy = new ProxyViewContainer();
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         outer.addChild(proxy);
         proxy.addChild(createBtn("1"));
 
@@ -45,7 +46,7 @@ export function test_children_registered_in_parent_grid_layout_on_attach() {
     var outer = new GridLayout();
     var proxy = new ProxyViewContainer();
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         proxy.addChild(createBtn("1"));
         outer.addChild(proxy);
 
@@ -59,7 +60,7 @@ export function test_add_children_to_detached_proxy() {
     var outer = new StackLayout();
     var proxy = new ProxyViewContainer();
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         outer.addChild(createBtn("1"));
 
         proxy.addChild(createBtn("2"));
@@ -88,7 +89,7 @@ export function test_remove_proxy() {
 
     outer.addChild(createBtn("5"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         assertNativeChildren(outer, ["1", "2", "3", "4", "5"]);
         outer.removeChild(proxy);
         assertNativeChildren(outer, ["1", "5"]);
@@ -111,7 +112,7 @@ export function test_remove_child_of_attached_proxy() {
 
     outer.addChild(createBtn("5"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         assertNativeChildren(outer, ["1", "2", "3", "4", "5"]);
         proxy.removeChild(testBtn);
         assertNativeChildren(outer, ["1", "2", "4", "5"]);
@@ -132,7 +133,7 @@ export function test_insert_inside_porxy() {
 
     outer.addChild(createBtn("5"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         assertNativeChildren(outer, ["1", "2", "4", "5"]);
         proxy.insertChild(createBtn("3"), 1);
         assertNativeChildren(outer, ["1", "2", "3", "4", "5"]);
@@ -152,7 +153,7 @@ export function test_insert_after_porxy() {
     proxy.addChild(createBtn("3"));
     proxy.addChild(createBtn("4"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         assertNativeChildren(outer, ["1", "2", "3", "4"]);
         outer.insertChild(createBtn("5"), 2);
         assertNativeChildren(outer, ["1", "2", "3", "4", "5"]);
@@ -170,7 +171,7 @@ export function test_proxy_does_not_stop_request_layout_bubble() {
     var btn = createBtn("2");
     proxy.addChild(btn);
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         assertNativeChildren(outer, ["1", "2"]);
         waitUntilElementLayoutIsValid(outer);
         TKUnit.assert(outer.isLayoutValid, "outer container isLayoutValid should be true");
@@ -185,7 +186,7 @@ export function test_proxy_iniside_page() {
     var proxy = new ProxyViewContainer();
     proxy.addChild(createBtn("1"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         var page = <Page>views[1];
         waitUntilElementLayoutIsValid(page);
     };
@@ -200,7 +201,7 @@ export function test_proxy_iniside_scroll_view() {
     var proxy = new ProxyViewContainer();
     proxy.addChild(createBtn("1"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         var page = <Page>views[1];
         waitUntilElementLayoutIsValid(page);
     };
@@ -215,7 +216,7 @@ export function test_proxy_iniside_border() {
     var proxy = new ProxyViewContainer();
     proxy.addChild(createBtn("1"));
 
-    function testAction(views: Array<viewModule.View>) {
+    function testAction(views: Array<View>) {
         var page = <Page>views[1];
         waitUntilElementLayoutIsValid(page);
     };
@@ -233,7 +234,7 @@ export function test_proxy_iniside_border() {
 //    items.push(new TabViewItem({ title: "tab with proxy", view: proxy }));
 //    tab.items = items;
     
-//    function testAction(views: Array<viewModule.View>) {
+//    function testAction(views: Array<View>) {
 //        var page = <Page>views[1];
 //        waitUntilElementLayoutIsValid(page);
 //    };
@@ -243,7 +244,7 @@ export function test_proxy_iniside_border() {
 
 // TODO: Proxy as a direct child to of ActionBar is not supported. Not sure if we want to support it.
 //export function test_proxy_iniside_actionBar() {
-//    function testAction(views: Array<viewModule.View>) {
+//    function testAction(views: Array<View>) {
 //        var page = <Page>views[1];
 //        var proxy = new ProxyViewContainer();
 //        proxy.addChild(createBtn("1"));
