@@ -1,9 +1,7 @@
-﻿import pageModule = require("ui/page");
-import button = require("ui/button");
+﻿import button = require("ui/button");
 import {DockLayout} from "ui/layouts/dock-layout";
 import TKUnit = require("../TKUnit");
 import helper = require("./layout-helper");
-import navHelper = require("../ui/helper");
 import testModule = require("../ui-test");
 import layoutHelper = require("./layout-helper");
 import commonTests = require("./common-layout-tests");
