@@ -3,7 +3,7 @@ import view = require("ui/core/view");
 import dts = require("ui/page");
 import styleScope = require("../styling/style-scope");
 import {ActionBar} from "ui/action-bar";
-import {DependencyObservable, PropertyMetadata, PropertyMetadataSettings, PropertyChangeData, Property, ValueSource} from "ui/core/dependency-observable";
+import {PropertyMetadataSettings, PropertyChangeData, Property, ValueSource} from "ui/core/dependency-observable";
 import * as style from "../styling/style";
 import * as fileSystemModule from "file-system";
 import * as frameModule from "ui/frame";
@@ -280,8 +280,11 @@ export class Page extends ContentView implements dts.Page {
 
     public _eachChildView(callback: (child: view.View) => boolean) {
         super._eachChildView(callback);
-
         callback(this.actionBar);
+    }
+
+    get _childrenCount(): number {
+        return (this.content ? 1 : 0) + (this.actionBar ? 1 : 0);
     }
 
     private _applyCss() {
