@@ -9,7 +9,6 @@ import proxy = require("ui/core/proxy");
 import color = require("color");
 import style = require("ui/styling/style");
 import font = require("ui/styling/font");
-import styling = require("ui/styling");
 import * as imageSourceModule from "image-source";
 
 var imageSource: typeof imageSourceModule;
@@ -191,7 +190,7 @@ export class TabView extends common.TabView {
     private _tabLayout: org.nativescript.widgets.TabLayout;
     private _viewPager: android.support.v4.view.ViewPager;
     private _pagerAdapter: android.support.v4.view.PagerAdapter;
-    private _androidViewId: number;
+    private _androidViewId: number = -1;
 
     private _pageChagedListener: android.support.v4.view.ViewPager.SimpleOnPageChangeListener;
 
@@ -227,7 +226,7 @@ export class TabView extends common.TabView {
         this._viewPager.setLayoutParams(lp);
         this._grid.addView(this._viewPager);
 
-        if (!this._androidViewId) {
+        if (this._androidViewId < 0) {
             this._androidViewId = android.view.View.generateViewId();
         }
         this._grid.setId(this._androidViewId);
